Relay: Configuring OSS Plugin for Legacy GraphQL
Summary: Adds support for configuring the `babel-relay-plugin` for use with the legacy version of GraphQL. (The goal of this project is to switch Facebook onto the same version of GraphQL as the open sourced.)

This adds the ability to configure:

- Name of the mutation input argument.
- Whether generated fields should use `camelCase` or `snake_case`.

This also changes the custom validator inteface so that `graphql` is an injected dependency. This allows the validator to use the identical instance of `graphql` that `babel-relay-plugin` uses (and not another instance of it, which causes problems with `instanceof` checks among other things).

Reviewed By: josephsavona

Differential Revision: D2684161

fb-gh-sync-id: c974538c90baa13c151d2aba5f73568b4a6acb14