Skip non-matched fragments
Summary:
We were looking for fields that were under fragments tht didn't match the type of the node, which was causing disk cache to have a lot of false negatives. This led to us not rendering early even when we might have had sufficient data from disk.
This diff allows us to skip traversing fragments that don't match the type of the current node

Reviewed By: josephsavona

Differential Revision: D2715044

fb-gh-sync-id: 663911b755473805124ba31ff5edc0c60a52ae68