Store __typename for all records
Summary: We originally added logic to store or infer the `__typename` of records for data synchronization purposes. For that use case the type wasn't needed for client records. However, now that we're using it to make operations such as writing, diffing, etc faster (by pruning traversal of non-matching fragment types), we should store the type for all records.

Reviewed By: yungsters

Differential Revision: D2714539

fb-gh-sync-id: 8788a5813e42e737691095c99eee2509558cc46b