Fixes #511 - Unable to use Relay with stateless components
Summary:
Resolves the React warnings seen when passing a function component to Relay.createContainer().
Adds a check to skip adding a "ref" to the container component if it is a function component.

This does mean setVariables() & forceUpdate() callbacks will not be provided the component
if you passed createContainer() a function component.

Please let me know if I am on the right track here or what needs more work. :)
Closes https://github.com/facebook/relay/pull/525

Reviewed By: josephsavona

Differential Revision: D2727708

fb-gh-sync-id: 70ee050e57efff668721867584a8c07bf53d954a