Update APIReference-Container.md typo
Summary:
In the current docs it says we should use `prepareVariables` and `initialVariables` as `static` variables of a class.

That doesn't work and is misleading, so I updated it to display it the way it actually works: As keys of an Object.

i.e.:
```js
// from
Relay.createContainer(Component, {
  static initialVariables = {...}
  ...
}

// to
Relay.createContainer(Component, {
  initialVariables: {...}
  ...
}
```

Hope it helps.
Closes https://github.com/facebook/relay/pull/645

Reviewed By: yungsters

Differential Revision: D2736851

fb-gh-sync-id: 2120b5e5df602a50f7f6be6c32c0369de048d318