Update www to jasmine2
Summary:
This is the big one. See D2728470 for an explanation.

I had to relax some checks because previously jasmine wouldn't check functions for equality (wtf!?).

I also patched jasmine not to be so strict about objects *for now*. The type of the object and undefined keys aren't being checked any more with this diff, same as in jasmine 1. Removing these patches results in about 800 test failures and I would like to clean this up later; this diff is already too big. I did start fixing these two failures in a lot of tests, so you'll see it in this diff, but then I gave up.

bypass-lint

Reviewed By: fabiomcosta

Differential Revision: D2728491

fb-gh-sync-id: ca956c96d59c6c644428f6c7ecf3802554b1acb6