Always write rootcall info into rootCallMap
Summary:
D2698788 intruduced a bug were we no longer write to nodeData if we have it in cachedData. This only fixes the missing writes in rootcallMap
We'll need a more complicated fix for rest of the writer, but we can always just write to the rootcall map since that happens really rarely.

Reviewed By: josephsavona

Differential Revision: D2743003

fb-gh-sync-id: c229932415eb7989ccdef2fc960e84e8fc83bab3