more unit test checks
Summary: josephsavona found that commenting out the line that writes to cached record didn't cause  any of the checks to fail other than a infinite loop error. The goal of readRelayDiskCache is to ensure data is avaliable to components. We should make sure this function actually reads data into cachedRecords. I added a bunch more checks to ensure the data we want is expected accessible in the RelayRecordStore.

Reviewed By: josephsavona

Differential Revision: D2765274

fb-gh-sync-id: 2e3335b2b257bda0bf71c04fd28100befcead430