handle repeated edges more gracefully
Summary:
I assumed that duplicated edges can only be returned across multiple request, and haven't really  handled this case very carefully. Apparently certain entqueries can be flaky and return dupped edges. We did dedup across requests but not within a request. Because it still don't really make sense for the same node to show up in multiple places in a connection, we handle this by not writing the second deuplicated edge.
The original code updated the max and min index without writing anything into that index, which caused the undefined error. I moved the duplication check higher so we don't increment the indices until we know this edge is not in the segment yet.

https://fb.facebook.com/groups/thefacebook/permalink/976164232432192/ for more context

Reviewed By: josephsavona

Differential Revision: D2787334

fb-gh-sync-id: a7935edb0c6df0bc2e16db0dc3a04930fdc12242