Relay: Fix Writing Deferred Fragments
Summary:
D2786199 (8885f235ab77af45699b2e7edc5abab23d2aee8e) changed `getFragmentID()` to throw on nodes that were cloned with new children to allow certain guarantees about the use of `getConcreteFragmentHash()`. Unfortunately, this broke one of two usages of `getFragmentID()`. The two usages of `getFragmentID()` are:

- Since D2791193 (a33c0b9d81fd973ff2ece8855b86a7b86fdfb0ac), `printRelayQuery` uses `getFragmentID()` to memoize the printer traversal.
- In `RelayContainer` and `RelayQueryWriter`, `getFragmentID()` is used to communicate the status of deferred fragments.

The latter of these two usages broke because when deferred fragments are used, fragments may be cloned to exclude the deferred fragments. To generalize, the problem is that `getFragmentID()` is indeed necessary for fragments that are cloned with new children.

This revision fixes the bug by changing the guarantees of `getFragmentID()` to return the same ID for two fragments with different children via cloning. This fixes the bug related to deferred fragments, but it allows `printRelayQuery` to continue working because `printRelayQuery` checks `hasConcreteFragmentHash()`.

Reviewed By: josephsavona

Differential Revision: D2792356

fb-gh-sync-id: b95555b7c01ea2cf2730cea9be2b2b1bb1a7f76f