Allow variable connection arguments
Relay Connections use a handful of arguments for pagination (first, last, before, after). These connection arguments are
statically validated by the babel-relay-plugin, and then checked again at runtime by GraphQLRange. However, if the
developer wants to use variables for the connection arguments (for example to support switching between `first` & `last`
at runtime), the validation performed by babel-relay-plugin gets in the way and prevents this, which forces the
developer to use some workarounds to "trick" the validation (see issue #677 for some examples).

This commit changes the babel-relay-plugin validation to allow otherwise invalid combinations (like `first`+`before`) as
long as all the arguments are variables. When the connection arguments are variables, babel-relay-plugin effectively
"trusts" that the developer will provide the correct arguments at runtime, and it becomes the job of GraphQLRange to
validate the calls given the actual values. GraphQLRange was already doing this validation, so I simply added another
unit test to verify that it does, in fact, reject the only unsupported runtime combination (`first`+`last`).

NOTE: babel-relay-plugin and GraphQLRange have different "expectations" for what combinations of arguments to a
connection are, in fact, valid. For example, babel-relay-plugin claims that `before`+`first` is invalid, even though
GraphQLRange happily accepts that combination. I chose not to change this, trusting that there was probably a good
reason for the discrepancy and that babel-relay-plugin was just being more defensive. However, I did notice that
babel-relay-plugin will (amusingly) allow the one combination of arguments that GraphQLRange cannot support:
`first`+`last`, so I added a new validation to babel-relay-plugin to prevent that... unless they are both variables, of
course!

This fixes issue #677.