Relay: Unique Hashes for Inline Fragments
Summary:
There is currently a bug with the concrete fragment hashes generated by `babel-relay-plugin`.

Each fragment is supposed to have a unique hash, but the input to this hash is the entire string content of the `Relay.QL` template literal. This means that each fragment (including inline fragments) will have the same concrete fragment hash. This causes bugs like not correctly printing two inline fragments (which uses the concrete fragment hash as a memoization key).

This fixes the bug by generating a unique concrete fragment hash for each fragment in a `Relay.QL` expression.

Reviewed By: steveluscher

Differential Revision: D2799942

fb-gh-sync-id: ef34db199d1ee8533277fd2b585d6a2743f00ef0