Relay: Rename Concrete Fragment Hash to Structural Hash
Summary:
Renaming "concrete fragment hash" to "structural hash" because it turns out that we //do// need what used to be called the "concrete fragment ID". That is, we need a hash that uniquely identifies each **instance** of a concrete fragment node (not just each fragment **structure**).

Since the term "concrete fragment hash" describes both of these, I'm going to avoid naming either of them with this term.

Reviewed By: yuzhi

Differential Revision: D2805427

fb-gh-sync-id: 3dc3bd653b328c7e080952de9650a34eaa0ceee2