Relay: Fix Disk + GC Race Condition
Summary:
When the disk cache and garbage collection are enabled, a race condition can occur.

== Context

Any record that exists in the record store (including `cachedRecords`) can be vended to containers and subcribed to. When containers subcribe to records, they increment the record's reference count in the garbage collector. All records must be registered in order to have their reference count incremented.

The function responsible for reading data out of the disk cache into the record store, `readRelayDiskCache`, incrementally (using a BFS-style traversal) creates records in `cachedRecords`, but those nodes do not currently get registered with the garbage collector until after the entire traversal is complete (via the change tracker).

== Problem

The result is that in the middle of a `readRelayDiskCache` traversal, if a container attempts to subscribe to a record that has not yet been registered with the garbage collector, `RelayGarbageCollector` fatals.

== Solution

This revision fixes the bug by ensuring that records created by `readRelayDiskCache` also atomically register the record with the garbage collector.

Reviewed By: yuzhi

Differential Revision: D2818759

fb-gh-sync-id: ab9ab922e922d1807e65556d26ea726f5c119d77