expose a way to read with a fragment
Summary:
Reading from set of queries makes sense in the context of GraphQLQueryRunner, but as we explore things like RelayConnection manager, being able to read with an dataID + fragment is going to be pretty useful. Currently in RelayConnectionManager, Joe has to hack a root query even though the dataIDs might be client ids. This should help make those less hacky

I didn't really change the logic, only rearranged some of the functions so we can reuse existing fuctions to read these fragments

Reviewed By: yungsters

Differential Revision: D2822684

fb-gh-sync-id: c8b4155cb94b6b03f088cf1feb4e72a3d509b5fc