Relay: Improve Fragment Pointer Detection Warning
Summary:
When a fragment pointer is set as the wrong prop, `RelayContainer` currently warns in `__DEV__`. However, the recent change of the fragment hash to a numeric string caused false positives to occur when a fragment prop is missing and another prop is assigned an array.

This fixes the bug by using the new `RelayRecord.isRecord` method.

Reviewed By: kassens

Differential Revision: D2816791

fb-gh-sync-id: 5677b27bd8b017b90112bfbb4797e00c6c56af5b