Relay: Use Index Paths for Serialization Keys
Summary:
This introduces `RelayQueryIndexPath` which tracks fragment indexes to the nearest parent field during query traversal. This replaces the existing logic used to generate field serialization keys, which is used exclusively for 1) printing queries and 2) writing the server payload into the store.

----

Currently, we encode fields with arguments by serializing all arguments and hashing them. This is problematic for two reasons:

- Cost of creating serialization keys increases with more arguments or more complex values.
- Creating the serialization key is dependent on the value of the argument values.

The latter prohibits any possibility of serializing query strings in a build step (e.g. to persist queries using IDs on the server).

Reviewed By: steveluscher

Differential Revision: D2748835

fb-gh-sync-id: 9356102126a4e0d656466bbebe339dcd5dcbaef6