Relay: Temporarily Remove Printer Memoization
Summary:
Disables the printer memoization for now. With the introduction of index paths as serialization keys, the memoization key is no longer valid and leads to fields being printed with incorrect serialization keys.

I plan to revamp the printer to get this performance optimization back. But for now, I want to unbreak master.

Reviewed By: yuzhi

Differential Revision: D2832503

fb-gh-sync-id: 635d7b1358a004e5e80ce86c92a1d2d00e772132