Don't try to setState after a fragment data update if component not mounted
Summary:
In certain cases, if a RelayContainer is mounted and unmounted quickly, a warning about setting state on an unmounted component is thrown.

This happens due to the fact that in the callback for fragment data updates from the GraphQLStoreQueryResolver, we `setState` without checking whether or not the component is mounted.

This PR resolves that issue.
Closes https://github.com/facebook/relay/pull/730

Reviewed By: yungsters

Differential Revision: D2817201

fb-gh-sync-id: 35eeefeb9affc1eb57d2cad22a9d97e82b4b9d55