Relay: Avoid Binding Functions in Query Traversal
Summary:
While using the Chrome Profiler to benchmark some other code, I found that the cost of binding some anonymous function was non-trivial. It turns out that you should not bind a function in a very hot code path.

This adds a `context` argument to some of the traversal methods so that we can avoid binding the function thousands of times per query.

Reviewed By: kassens

Differential Revision: D2848579

fb-gh-sync-id: d7ce136f99b582a3be9c10dc733a2242265198a7