New writer so we know where we are writing to instead of implicitly falling through the stores [1/N]
Summary:
There has been a few instances where we thought we were writing to one store but was actually writing to another store (including a SEV).
This aims to seperate the writing functions from the nice convinience of RelayRecordStore, which does the fall through reading of our stores, so we can
be more aware of what we are actually writing to.
- Copied over RelayRecordStore and replaced `Store` with `Writer`. There are some read functions I might need for the write path, so I am only removing functions I know we don't need for now. Will clean up more once the current writers are switched over and I can be more sure that they are not needed.
- Removed the extra `records` and `rootCallMaps`
- Added `isOptimisticWrite` to the constructor so we know if we are dealing with an optimistic mutation or not
- Removed `hasOptimisticUpdates`, `getClientMutationIDs` since those seem to be purely used by RelayStoreData and don't need a separate writer function
Next steps: convert the writers to use this new RelayRecordWriter

Reviewed By: yungsters

Differential Revision: D2851115

fb-gh-sync-id: 384f80294aa80e7571da4f50966f11610777ad74