Make babel-relay-plugin run before other transforms
Summary:
`babel-relay-plugin` depends transforms the “Relay.QL” TaggedTemplateExpression nodes in the source. If it does not run before the “es2015-template-literals” transform, it will not transform the Relay.QL template literals correctly.

This is important, because with Babel 6, you can’t control the plugin order. And so in a case like React Native, where plugins from React Native’s babelrc are loaded before the projects babelrc, it’s impossible to use the Babel Relay Plugin without overriding the entire transform list.

Note - I made this change in the babelAdapter, rather than in the plugin code itself, in order to retain Babel 5 compatibility.
Closes https://github.com/facebook/relay/pull/714

Reviewed By: yungsters

Differential Revision: D2862250

fb-gh-sync-id: 1c7342bda81b4dfbb7746ce1edee110ecd7b759b