Manually test fields inferred from null payload.
Summary:
Let's say you get a payload like this:

    {address: null}

You have no information to infer whether this field is scalar or not.

This diff switches out our `toEqualFields` matcher for a manual test, because once we start to properly identify the `address` field as one that can have sub-selections, that matcher will start to strip it out because it's empty (the matcher uses `clone()` as part of its normal operation, which strips out empty non-scalar fields).

Reviewed By: yungsters

Differential Revision: D2863273

fb-gh-sync-id: b897803a3b1c981dc045b2abb4caaa69737cce6f