Replace `isScalar` with `canHaveSubselections`
Summary: Now that we print `canHaveSubselections` metadata, we can get rid of `isScalar` in favor of the more rigorous `canHaveSubselections`. Previously, `isScalar` made the assumption that a field was a scalar if it had no children. After this diff, the definition of a field that can have subselections is one that's not of type `GraphQLScalar` or `GraphQLEnum`.

Reviewed By: josephsavona

Differential Revision: D2863892

fb-gh-sync-id: 5c1a7e635cf0eabf37668bfbdfba17e797b9a835