Remove two more unnecessary Flow casts
Summary:
Depends on D2875226.

(And separate from it because I am not sure I fully trust Flow on this one. I
tried removing this yesterday and was getting the same error mentioned in the
comment; admittedly I had a couple of in-progress changes conflated in my
worktree so perhaps it is ok after all. In any case, the CI will tell all.)

Reviewed By: josephsavona, yungsters

Differential Revision: D2875305

fb-gh-sync-id: a1bc4ceeb28ad0c931d3dfbeac5265fa5b33dbd0