Preserve empty non-leaf nodes in fat queries
Summary:
The recent change to [[https://github.com/facebook/relay/commit/c51db5cc3a434ead82317c24aae9b21b2bac1716 | replace `isScalar` with `canHaveSubselections`]] introduced a bug with mutations where "pattern" nodes in fat queries are inadvertently removed. In the following fat query:

```
fragment on Comment {
  doesViewerLike # scalar, no possible subselections
  likers # non-scalar, can have subselections
}
```

The `likers` field was being removed when fat queries are flattened in `RelayMutationQueue`.

- added a `preserveEmptyNodes` option to `flattenRelayQuery` that, when enabled, will retain any empty non-scalar nodes
- changed `RelayMutationQueue` to pass the new option if the fat query fragment has the `relay(pattern: true)` directive

The concrete type definition & query builder are also updated to reflect the presence of the `pattern` metadata that is set when the `relay(...)` directive is present.

Reviewed By: steveluscher

Differential Revision: D2881040

fb-gh-sync-id: dd7d06f3e3345b270acc9d83349711d86b533f68