Avoid collisions storing optimistic range data on records
Summary:
We were storing optimistic range updates directly on the record with keys
"append", "prepend" or "remove". This could collide with user-supplied fields
of the same name, so we should instead follow the practice we've already
established for the other metadata fields, which is putting then in
`__underscorizedForm__`.

Included changes in `GraphQLRange`:

- `optimisticData` is now `queuedRecord`, which is more descriptive, and
  consistent with the terminology we use at the call site.
- `deleteIDs` is now `removeIDs` for consistency with the terminology from
  `GraphQLMutatorConstants`.
- `let`- and `const`-ify; not the entire file but the lines that I was going to
  touch anyway.
- Avoid some unnecessary array allocations; make helper accessors instead.

Reviewed By: yungsters

Differential Revision: D2880739

fb-gh-sync-id: e812b3f89fc1b4b578f58327d4f069a52cd28e1d