Sync RelayRecord metadata keys with Record Flow type
Summary:
Adds the missing keys (`__mutationIDs__` and
`__resolvedDeferredFragments__`) and moves the associated `Record` and
`RecordMap` types into `RelayRecord`, right next to the metadata map, so that
these two will be less likely to get out of sync in the future.

Note that there are a couple of circular dependencies introduced here, but only
in Flow `import type` declarations, and Flow can handle those just fine.

- `Record` depends on `RelayQueryPath`, which depends on `RelayRecord`.
- `Record` also depends on `GraphQLRange`, which depends on `RelayRecord` too.

Reviewed By: yungsters

Differential Revision: D2887118

fb-gh-sync-id: 962fecdfeeb4a06e962ba90d929dc668df89aadb