Relay: Smarter Plugin Cache Breaking
Summary: Changes the `babel-relay-plugin` build script to output a `lib/HASH` file that servers with caching (e.g. React Native) can use as a cache breaker (instead of having to check every file in the plugin). This is especially valuable when developing new plugin features because it enables testing without manually dirtying the package version number.

Reviewed By: josephsavona

Differential Revision: D2888573

fb-gh-sync-id: 1e0f4f3c4c24d9bf02e54c21cecf54ef8b780ea1