Add RelayRecord.createWithFields and use it
Summary:
For the create-then-populate-with-fields use case.

Note there is one place in `GraphQLFragmentPointer` where we were using it and
were stretching the types a little bit; this is perhaps evidence that we
shouldn't actually be using `RelayRecord` here, as what we are returning is
"recordish", but not really a record.

I've changed the types here to be more truthful about what we're actually
returning, but I'm still not totally happy with it.

Reviewed By: yungsters

Differential Revision: D2897705

fb-gh-sync-id: 19e7299ec8e7dd5f1b60d88d45856dd0c29f5057