Deprecate RelayRenderer `Component` prop in favor of `Container`
Summary:
The code comments in `RelayRenderer` mention a `Container` prop,
suggesting that this is what we were intending to use at some point during the
design of the API, but as implemented the prop is actually `Component`.

As evidenced in this GitHub issue, this may cause confusion and there's an
opportunity for us to make things clearer by changing the name here:

https://github.com/facebook/relay/issues/785

This diff continues to support `Component`, but issues a warning in `__DEV__`.
The code to handle the deprecation is ugly but fortunately can be short-lived.
As `RelayRenderer` isn't publicly documented yet, we don't have to keep the code
to support the deprecation around for long; I'm thinking after the next release
we can rip it out, with plenty of time to update internal callers.

Reviewed By: yungsters

Differential Revision: D2901998

fb-gh-sync-id: dfc7a44babf1f6a43749691d3430e92884c94454