Downgrade invariant to warn in GraphQLQueryRunner
Summary:
This invariant was added long before we had disk cache. We expected no more state change after an error or done except abort. However in the case of airplane mode and disk cache, we can read data from diskcache after we receieved a network error, so the invariant is no longer true.
There is a task to revisit onReadyStateChange with task so this is just the temp fix

Reviewed By: wincent

Differential Revision: D2903180

fb-gh-sync-id: cfca673a39550d36bf952480ce61e99be10e3053