prevent gc conflicts during data-fetching/rendering
Summary:
This is a short-term fix for some bugs discovered during test of GC. The general problem is that GC will by definition remove records that are unreferenced, but this can cause required data to be removed between the start of rendering a `RelayRenderer` and when its container tree is actually rendered (ie, data could be evicted between being written to the store and referenced by a component).

The long-term fix will be to ensure that any data that *may* be needed by a container tree is referenced immediately (ie, upon diffing or writing to the store). The short term fix (this diff) is:
- Prevent GC from running while there are mounted RelayRenderers.
- Delay GC execution via `setImmediate` to handle cases where one renderer is unmounted and another remounted within the same event loop.
- Ensure that reference counts are not added for nodes that do not exist in the store (this is related bug: those nodes cannot have been registered with GC)

Reviewed By: yungsters

Differential Revision: D2903161

fb-gh-sync-id: 0eee09cfae065b95e69327c257706a77ff44a21a