Add USERS.md
Summary:
As discussed at our recent chat, we'd like to have a "Users" page that
people can add themselves to. We're not sure that keeping it in the repo like
this is the right thing (as opposed to, say, a wiki page), but we'd like to try
the pull request model and have the community build this up. If it turns out to
be suboptimal we can always come up with a new plan.

Reviewed By: elynde

Differential Revision: D2907962

fb-gh-sync-id: 412983a48aba6a5736d7085ab864c568a1f58d77