Add AlphaSights to the list of Relay users
Summary:
Hey Relay team, thanks for open sourcing Relay and supporting it as a product :)

We started using React/Flux in version 2 of our internal recruitment application (Rails backend) and have been hoping that Relay will help solve the problems around data loading.

We've been using it in production in by building new features with it and slowly rewriting the other features. So far so good, the graphql-ruby gem and Relay has worked seamlessly for us.
Closes https://github.com/facebook/relay/pull/800

Reviewed By: yungsters

Differential Revision: D2912057

fb-gh-sync-id: ad0de2d4ca5e60050bca3d23af772690f209aa8c