Track fragments on null linked fields
Summary:
This fixes #745.

Essentially, what's happening is that the query tracker isn't being updated by `diffRelayQuery` when the query is on a node that is currently `null` in the local store. This causes Relay to lose track of what queries the components are expecting and therefore future mutations form incorrect queries.

My current fix just changes a bit of logic in `diffRelayQuery` to always track null linked fields. I'm not sure it's correct though - I'd like to write some code that ensures the `RelayQueryTracker` correctly merges the two tracked queries that `diffRelayQuery` registers via `trackNodeForID`, but maybe that's already covered by existing tests.
Closes https://github.com/facebook/relay/pull/774

Reviewed By: yuzhi

Differential Revision: D2901088

fb-gh-sync-id: 27d7416f39efb775fba3249579474c70146791f9
shipit-source-id: 27d7416f39efb775fba3249579474c70146791f9