Account for duplicate variables in printRelayQuery
Summary:
This fixes #784 on github, in which Relay can sometimes print queries such as:

```
query ($orderBy: _GroupOrdering, $orderBy2: _GroupOrdering) {
  viewer {
    allGroups_<hash>: allGroups(orderBy: $orderBy) { count }
    allGroups_<hash>: allGroups(orderBy: $orderBy2) { count }
  }
}
```

Which does not pass static validation, since the same alias is used twice with different arguments. Although in this case the values of both variables will be the same (otherwise the alias would differ), GraphQL's static validation pass does not know this and rejects the query. The validation is static to allow things like persisted queries.

The solution is to avoid printing a new variable if one already exists with the same value. The original implementation iterated over all variables to find any whose value matched, but I asked for the OP to change to this to a `Map`. After importing I realized that input object values weren't handled correctly so I ended up modifying this and adding an extra test case.

-- original pr --

This pull request relates to #784. I've adjusted the `createVariable` function, within `src\traversal\printRelayOSSQuery.js`, so that it first checks the existing variables before creating a new variable. This is to assist Relay when printing queries that contain overlapping arguments. (If the variable type and value is the same, do we need to continue creating variables with the same type and value? This PR adds the logic that answers that with a *no*.)

Without knowing what existing code is at my disposal within all of the FB libraries, I went ahead and just use a simple for loop over the existing keys, since the return state needs the key anyway.

I also created a test case that, prior to adjusting `createVariable`, failed. After adjusting `createVariable` the test passes, and no other existing tests seem to fail. Yay?
Closes https://github.com/facebook/relay/pull/801

Reviewed By: steveluscher

Differential Revision: D2913333

fb-gh-sync-id: 6b7dbed2ce30254a7922a3e3c9614cf1aed1087f
shipit-source-id: 6b7dbed2ce30254a7922a3e3c9614cf1aed1087f