add metadata regarding if any data is missing
Summary:
This adds a way for product code to figure out if the RelayStore might be missing data for a particular container with out needing custom logic to manually walk through all the supplied props. Product code will be able to check this.props.relay.hasPartialData(<object>) to determine if data is missing and might need an forceFetch or primeCache.
- Adds a new bit for whether a component might have partial data into the RelayRecordStatusMap
- Updates readRelayQueryData to keep track of state when we run into missing data and write it into the status
- Add a function in RelayContainer to check if a subtree of props might be missing data in the store.

Reviewed By: josephsavona

Differential Revision: D2915046

fb-gh-sync-id: c0738027a191dd9cb2e537a8cf03477cdd2ff540
shipit-source-id: c0738027a191dd9cb2e537a8cf03477cdd2ff540