infer fragment name from prop
Summary:
This searches parent nodes of RelayQL templates for a property the template is keyed to, and uses that property name as part of name of anonymous fragments ("fragment on Type"). The reason is that developers referencing a Relay fragment from a native query must do so using the property the Relay fragment is keyed to. It won't suffice for developers to specify their own invented fragment names, so this diff effectually autogens the correct fragment name on their behalf. While the purpose here is just for Relay Native, I've updated all fragment names since the current fragment name (the document name) appears to be unimportant as it is.

I'm not super happy with this solution since such a property isn't guaranteed to exist. But, in order to resolve a RelayQL expression from native, we unfortunately have to know the prop name. (For static query generation we already enforce that developers write their RelayQL expressions as `fragments: { key: ()=>RelayQL``, }` anyway.)

Reviewed By: josephsavona

Differential Revision: D2917529

fb-gh-sync-id: cace2e5e57ae50d1a153e9e691aece0f6d479eef
shipit-source-id: cace2e5e57ae50d1a153e9e691aece0f6d479eef