Explicitly set all metadata when building queries
Summary:
More comprehensive version of the fix at D2895158 - refetch queries (`setVariables` and `forceFetch`) were not properly requesting the `__typename` of the record, which was therefore being written with `Node` as the type. The reader would then find that the type did not match and return empty data to the component.

This changes `RelayQueryRoot.build` to require a metadata object with all keys present, and updates all call sites to explicitly set the correct metadata (and adds regression tests for each).

Closes https://github.com/facebook/relay/pull/792

Reviewed By: steveluscher

Differential Revision: D2906348

fb-gh-sync-id: fd7027b66e3708a3f19f2dcf0a1b35ccfa3126a9
shipit-source-id: fd7027b66e3708a3f19f2dcf0a1b35ccfa3126a9