Replace $Enum with $Keys
Summary:
steveluscher pointed out in D2917170 that we might want to replace
our use of `$Keys` with string unions. I wanted to see how much we're using
`$Enum`/`$Keys` to see how many of those sites would be amenable to this
treatment, and noticed that we're not even using `$Keys` consistently (it used
to be `$Enum`, but was renamed a while back).

Lets make the names consistent, at least, for now, and tackle the string union
thing as another step.

Reviewed By: josephsavona

Differential Revision: D2917827

fb-gh-sync-id: 934e11b0745bc99a901cec1ef7695e876f8ab8eb
shipit-source-id: 934e11b0745bc99a901cec1ef7695e876f8ab8eb