make hasPartialData accessible by children
Summary: `hasPartialData` was added in a previous commit. This exposes the function to children of RelayContainer. If somehow an server updates remaps a node and cause it to no longer have all the fields needed by the component, `hasPartialData` will return `true`. Client can choose to refetch the data if that happens.

Reviewed By: josephsavona

Differential Revision: D2930667

fb-gh-sync-id: c71b0e28c675dd9bcf6abeb2e88027d7b7b0d390
shipit-source-id: c71b0e28c675dd9bcf6abeb2e88027d7b7b0d390