Add testEditDistance
Summary:I'll be using this up next to provide helpful suggestions when bad config is supplied in a mutation config. This will allow us to detect typos in key names (eg `fieldIDS` instead of `fieldIDs`).

Additionally, we could consider using it to provide helpful feedback in other places, including in the printer and elsewhere.

The edit distance algorithm employed comes from Berghel and Roach, and it is chosen because it is faster than the typical Levenshtein dynamic-programming approach, and is specifically well-suited to our use case here of detecting likely fuzzy matches and aborting once a certain distance threshold is achieved.

Reviewed By: yungsters

Differential Revision: D2934725

fb-gh-sync-id: 1df147f2ebbec06e80ed1cdce15586a1e2ae8b80
shipit-source-id: 1df147f2ebbec06e80ed1cdce15586a1e2ae8b80