Teach validateMutationConfig about optional keys
Summary:As per my comment in D2934793, I've done an audit trying to determine which keys in mutation config are required and which are optional. For now, it looks like only `parentName` on `RANGE_ADD` is optional, although I will loop back and double check as this was just on the basis of a quick scan.

In any case, this diff makes `validateMutationConfig` aware of the possibility that properties may be optional, and adjusts the validation accordingly.

Reviewed By: yungsters

Differential Revision: D2935376

fb-gh-sync-id: d24ace787c21f77b5d68c1c50be0058f0e858830
shipit-source-id: d24ace787c21f77b5d68c1c50be0058f0e858830