Move generateRQLFieldAlias out of "legacy" folder
Summary:This diff moves `generateRQLFieldAlias` module out of the "legacy/store" folder and into the non-legacy "query" folder (because it has more to do with queries than stores, and although we use this in a bunch of different places in the test suite, the only non-test use is inside `RelayQuery`).

This is a straight move, apart from adding `flow` annotation and the corresponding types.

Reviewed By: yuzhi

Differential Revision: D2941788

fb-gh-sync-id: 3b61bec54948d40d2abc6977a625996d028a5b5c
shipit-source-id: 3b61bec54948d40d2abc6977a625996d028a5b5c