Create build-time fragment ids to identify fragments across threads
Summary: RelayContainer uses fragment pointers - opaque metadata on fragment props - to verify that data for a component was fetched by its parent. The current implementation relies on a runtime-generated id, however, this is limiting if we want to perform data-fetching in a separate thread. To do so with runtime generated IDs would require serializing the query itself across threads which imposes a runtime cost. Instead, we can generate a unique id for each fragment within the source code and use that to identify fragments and their results across threads. These new ids are a hash of the filename, full file text, and line/column number - effectively uniquely identifying the fragment's place structurally within the codebase (the closest static equivalent to "this exact runtime object").

Reviewed By: yungsters

Differential Revision: D2943010

fb-gh-sync-id: 179914e1adb3757db989777c37c895f276c2f96e
shipit-source-id: 179914e1adb3757db989777c37c895f276c2f96e