Make RelayRenderer run queries after mount
Summary:...to make sure `RelayRenderer` does not run queries during synchronous server-side rendering.

In the `RelayRenderer` tests, I had to replace `ShallowRenderer.render` with the traditional `ReactDOM.render`, because `ShallowRenderer.render` does not invoke some component lifecycle methods including `componentDidMount` that now run queries. But I did not change the logic of the tests, as **visible behaviour of `RelayRenderer` was fully retained.**

Added a test to check that `RelayRender` does not run queries on the server.

This PR is a forerunner of the PR that will make `RelayRenderer` isomorphic (see #589). That next PR will be pretty small and easily comprehensible, because this PR already did much of the necessary dirty work.
Closes https://github.com/facebook/relay/pull/625

Reviewed By: josephsavona

Differential Revision: D2727748

fb-gh-sync-id: 202cc772d15661532afd5eee6ae647bb03af9dcc
shipit-source-id: 202cc772d15661532afd5eee6ae647bb03af9dcc