Generate new concrete fragment ids on clone()
Summary:`RelayQueryFragment#isCloned` was added as a way to avoid relying on a fragment's `getCompositeHash` if it had been cloned - the idea being that a cloned fragment's hash could be identical to the pre-cloned fragment's hash since they shared a common concrete node.

Building on D2943010, which creates truly unique fragment IDs on static fragments and D2947985 which creates unique ids for dynamically generated fragments, we can now remove `isCloned` and change all call sites to just use `getCompositeHash` on all fragments regardless of being cloned or not.

Reviewed By: yungsters

Differential Revision: D2949393

fb-gh-sync-id: f23163d50bfbad79e8ab83b18f9398eff580cd2d
shipit-source-id: f23163d50bfbad79e8ab83b18f9398eff580cd2d